requirements.txt
Pillow>=10.0.0
imagehash>=4.3.1
tqdm>=4.66.0
click>=8.1.0
